+++ apps/web/src/eval/features/evaluation-extraction-datasets/evaluation-extraction-datasets.slice.ts
@@ -23,7 +23,6 @@ interface State {
   fileColumns: AsyncData<EvaluationExtractionDatasetFileColumn[]>
   uploader: UploaderState
   isUpdatingDataset: boolean
-  isRenamingDataset: boolean
 }
 
 const initialState: State = {
@@ -38,7 +37,6 @@ const initialState: State = {
     errors: null,
   },
   isUpdatingDataset: false,
-  isRenamingDataset: false,
 }
 
 const slice = createSlice({
@@ -159,17 +157,6 @@ const slice = createSlice({
       })
       .addCase(evaluationExtractionDatasetsThunks.updateOne.rejected, (state) => {
         state.isUpdatingDataset = false
-      })
-
-    builder
-      .addCase(evaluationExtractionDatasetsThunks.renameOne.pending, (state) => {
-        state.isRenamingDataset = true
-      })
-      .addCase(evaluationExtractionDatasetsThunks.renameOne.fulfilled, (state) => {
-        state.isRenamingDataset = false
-      })
-      .addCase(evaluationExtractionDatasetsThunks.renameOne.rejected, (state) => {
-        state.isRenamingDataset = false
       })
   },
 })